# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "clinicPlatform API",
    "version": "1.0.0",
    "description": "Auto-generated OpenAPI from Next app routes"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/api/roles": {
      "get": {
        "summary": "GET /api/roles",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Role"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "POST /api/roles",
        "responses": {
          "200": {
            "description": "Success"
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Role"
              }
            }
          }
        }
      }
    },
    "/api/services/{id}": {
      "patch": {
        "summary": "PATCH /api/services/{id}",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": ""
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "clinicId": {
                    "type": "string"
                  },
                  "clinic": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "duration": {
                    "type": "string"
                  },
                  "price": {
                    "type": "string"
                  },
                  "active": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "PUT /api/services/{id}",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": ""
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "clinicId": {
                    "type": "string"
                  },
                  "clinic": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "duration": {
                    "type": "string"
                  },
                  "price": {
                    "type": "string"
                  },
                  "active": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "DELETE /api/services/{id}",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": ""
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "clinicId": {
                    "type": "string"
                  },
                  "clinic": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "duration": {
                    "type": "string"
                  },
                  "price": {
                    "type": "string"
                  },
                  "active": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/services": {
      "get": {
        "summary": "GET /api/services",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Service"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "activeOnly",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clinicId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "summary": "POST /api/services",
        "responses": {
          "200": {
            "description": "Success"
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": [
            "name",
            "duration"
          ],
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "duration": {
                    "type": "string"
                  },
                  "price": {
                    "type": "string"
                  },
                  "active": {
                    "type": "string"
                  }
                },
                "required": [
                  "name",
                  "duration"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "name": "activeOnly",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clinicId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "emailVerified": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "image": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "clinicId": {
            "type": "string"
          },
          "clinic": {
            "$ref": "#/components/schemas/Clinic"
          },
          "patient": {
            "$ref": "#/components/schemas/Patient"
          },
          "appointments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Appointment"
            }
          },
          "consultations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Consultation"
            }
          },
          "prescriptions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Prescription"
            }
          },
          "accounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Account"
            }
          },
          "sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Session"
            }
          },
          "notifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Notification"
            }
          },
          "auditLogs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditLog"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "email",
          "password",
          "appointments",
          "consultations",
          "prescriptions",
          "accounts",
          "sessions",
          "notifications",
          "auditLogs",
          "updatedAt"
        ]
      },
      "Clinic": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "logo": {
            "type": "string"
          },
          "settings": {
            "type": "object"
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          },
          "patients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Patient"
            }
          },
          "appointments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Appointment"
            }
          },
          "services": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Service"
            }
          },
          "invoices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Invoice"
            }
          },
          "notifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Notification"
            }
          },
          "auditLogs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditLog"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "name",
          "users",
          "patients",
          "appointments",
          "services",
          "invoices",
          "notifications",
          "auditLogs",
          "updatedAt"
        ]
      },
      "Patient": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date-time"
          },
          "gender": {
            "type": "string"
          },
          "bloodType": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "emergencyContact": {
            "type": "string"
          },
          "emergencyPhone": {
            "type": "string"
          },
          "medicalHistory": {
            "type": "string"
          },
          "allergies": {
            "type": "string"
          },
          "currentMedications": {
            "type": "string"
          },
          "clinicId": {
            "type": "string"
          },
          "clinic": {
            "$ref": "#/components/schemas/Clinic"
          },
          "appointments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Appointment"
            }
          },
          "consultations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Consultation"
            }
          },
          "prescriptions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Prescription"
            }
          },
          "invoices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Invoice"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "dateOfBirth",
          "phone",
          "clinicId",
          "clinic",
          "appointments",
          "consultations",
          "prescriptions",
          "invoices",
          "updatedAt"
        ]
      },
      "Service": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "duration": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "float"
          },
          "active": {
            "type": "boolean"
          },
          "clinicId": {
            "type": "string"
          },
          "clinic": {
            "$ref": "#/components/schemas/Clinic"
          },
          "appointments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Appointment"
            }
          },
          "invoiceItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceItem"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "name",
          "duration",
          "price",
          "appointments",
          "invoiceItems",
          "updatedAt"
        ]
      },
      "Role": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RolePermission"
            }
          },
          "staffCount": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "name",
          "permissions",
          "updatedAt"
        ]
      },
      "Permission": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RolePermission"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "name",
          "label",
          "category",
          "roles",
          "updatedAt"
        ]
      },
      "RolePermission": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "roleId": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "permissionId": {
            "type": "string"
          },
          "permission": {
            "$ref": "#/components/schemas/Permission"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "@@unique([roleId,": {
            "$ref": "#/components/schemas/permissionId])"
          }
        },
        "required": [
          "roleId",
          "role",
          "permissionId",
          "permission",
          "@@unique([roleId,"
        ]
      },
      "Appointment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/AppointmentStatus"
          },
          "notes": {
            "type": "string"
          },
          "patientId": {
            "type": "string"
          },
          "patient": {
            "$ref": "#/components/schemas/Patient"
          },
          "doctorId": {
            "type": "string"
          },
          "doctor": {
            "$ref": "#/components/schemas/User"
          },
          "serviceId": {
            "type": "string"
          },
          "service": {
            "$ref": "#/components/schemas/Service"
          },
          "clinicId": {
            "type": "string"
          },
          "clinic": {
            "$ref": "#/components/schemas/Clinic"
          },
          "consultation": {
            "$ref": "#/components/schemas/Consultation"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "startTime",
          "endTime",
          "patientId",
          "patient",
          "doctorId",
          "doctor",
          "clinicId",
          "clinic",
          "updatedAt"
        ]
      },
...
      "InvoiceItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "unitPrice": {
            "type": "number",
            "format": "float"
          },
          "totalPrice": {
            "type": "number",
            "format": "float"
          },
          "invoiceId": {
            "type": "string"
          },
          "invoice": {
            "$ref": "#/components/schemas/Invoice"
          },
          "serviceId": {
            "type": "string"
          },
          "service": {
            "$ref": "#/components/schemas/Service"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "description",
          "unitPrice",
          "totalPrice",
          "invoiceId",
          "invoice"
        ]
      },
      "Notification": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "read": {
            "type": "boolean"
          },
          "userId": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "clinicId": {
            "type": "string"
          },
          "clinic": {
            "$ref": "#/components/schemas/Clinic"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "type",
          "title",
          "message",
          "userId",
          "user",
          "updatedAt"
        ]
      },
...
    }
  }
}